backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.54.0
python-dotenv==1.0.0
httpx>=0.23.0,<0.28.0
